refactored to use decimal instead of double for monetary values and quantities, rounding to 2 decimal places

+++ csharp/SupermarketReceipt.NUnit.Test/ReceiptPrinter.cs
@@ -84,7 +84,7 @@ namespace SupermarketReceipt
             return line.ToString();
         }
 
-        private string PrintPrice(double price)
+        private string PrintPrice(decimal price)
         {
             return price.ToString("N2", Culture);
         }
