+++ csharp/SupermarketReceipt/Offer.cs
@@ -12,7 +12,7 @@ namespace SupermarketReceipt
     {
         private Product _product;
 
-        public Offer(SpecialOfferType offerType, Product product, double argument)
+        public Offer(SpecialOfferType offerType, Product product, decimal argument)
         {
             OfferType = offerType;
             Argument = argument;
@@ -20,6 +20,6 @@ namespace SupermarketReceipt
         }
 
         public SpecialOfferType OfferType { get; }
-        public double Argument { get; }
+        public decimal Argument { get; }
     }
 }
